requirements.txt
APScheduler==3.10.4
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.1.2
Automat==22.10.0
backports.zoneinfo==0.2.1
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==41.0.5
daphne==4.0.0
Django==4.2.7
django-cors-headers==4.3.0
django-filter==23.5
django-paypal==2.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
drf-writable-nested==0.7.0
et-xmlfile==1.1.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
jwt==1.3.1
msgpack==1.0.7
mysqlclient==2.2.0
numpy==1.24.4
openpyxl==3.1.2
pandas==2.0.3
Pillow==10.1.0
priority==1.3.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==24.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.4
stripe==8.1.0
Twisted==23.10.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
websocket-client==1.7.0
zoomus==1.2.1
zope.interface==6.1
